error handling and working "create user"

+++ src/utils/cloudinary.js
@@ -1,27 +1,34 @@
-import { v2 as cloudinary } from "cloudinary"
-import fs from "fs"
+import { v2 as cloudinary } from "cloudinary";
+import fs from "fs";
 
 cloudinary.config({
   cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
   api_key: process.env.CLOUDINARY_API_KEY,
   api_secret: process.env.CLOUDINARY_API_SECRET,
-})
+});
 
-const uploadonCloudinary = async (localFilePath) => {
+const uploadOnCloudinary = async (localFilePath) => {
   try {
-    if (!filePath) return
+    if (!localFilePath) {
+      // console.log("File not found !");
+      return;
+    }
     //upload file on cloudinary
     const response = await cloudinary.uploader.upload(localFilePath, {
       resource_type: "auto",
-    })
-    //file has been uploaded successfully
-    console.log("File is uploaded on cloudinary successfully!! ", response.url)
-    return response
-  } catch (error) {
-      //remove the locally saved temp file as the upload operation got failed.
-      fs.unlinkSync(localFilePath)
-    return null
-  }
-}
+    });
 
-export { uploadonCloudinary }
+    //file has been uploaded successfully
+    // console.log("File is uploaded on cloudinary successfully!! ", response.url);
+
+    fs.unlinkSync(localFilePath);
+    return response;
+  } catch (error) {
+    //remove the locally saved temp file as the upload operation got failed.
+    
+    fs.unlinkSync(localFilePath);
+    return null;
+  }
+};
+
+export { uploadOnCloudinary };
